requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
fonttools==4.53.1
idna==3.7
importlib_resources==6.4.3
jmespath==0.10.0
kiwisolver==1.4.5
matplotlib==3.9.2
nls @ file:///Users/mian000/www/projects/alibabacloud-nls-python-sdk
numpy==2.0.1
oss2==2.18.6
packaging==24.1
pillow==10.4.0
PyAudio==0.2.14
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Wave==0.0.2
zipp==3.20.0
